Change drag threshold

+++ noteblock.h
@@ -23,7 +23,7 @@ public:
 signals:
     void noteDeleted(NoteBlock* o);
     void trySwap(NoteBlock* o);
-    void dragProgress(bool isVertical, float progress, NoteBlock* o);
+    void dragProgress(bool isVertical, qreal progress, NoteBlock* o);
 
 private:
     Ui::NoteBlock *ui;
